requirements.txt
beautifulsoup4
google-genai
pandas
pydantic
python-dotenv
PyMuPDF
requests
streamlit
urllib3
